Separate files and logic
This commit removes the logic to generate a specific playlist to an example folder, so this project looks more like an actual package that can be used by others (which is the goal of the project). This project should merely provide a way to generate playlists, but it should not contain specific logic, like selectors that can be used for generating playlists.